ModelServer/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
dill==0.3.7
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.47.0
fsspec==2023.12.2
humanfriendly==10.0
idna==3.6
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
onnx==1.15.0
onnxruntime==1.16.3
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
protobuf==4.25.1
psutil==5.9.7
py-cpuinfo==9.0.0
pyparsing==3.1.1
pyreadline3==3.4.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
sympy==1.12
thop==0.1.1.post2209072238
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
ultralytics==8.0.228
urllib3==2.1.0
pika==1.3.2
